requirements.txt
Flask
pymupdf
pdf2docx
pytesseract
Pillow
